cubane/requirements/common.txt
psycopg2==2.6.2
beautifulsoup4==4.5.3
lxml==3.7.2
django-htmlmin==0.9.0
Wand==0.4.4
requests==2.12.5
requests[security]
pycrypto==2.6.1
oauth2==1.9.0.post1
mailsnake==1.6.4
ipaddr==2.1.11
ifaddr==0.1.4
pydns==2.3.6
pyspf==2.0.12t
chardet==2.3.0
stripe==1.46.0
pyBarcode==0.7
idna==2.5